Chapter 8
Introduction to greedy algorithms, which is a way
to find the optimium solution locally and hope it
give an optimium, they are easy to write, this
makes them good approximation algorithm to solve
the NP-COmplete problems(e.g. Travelling Salesperson,
and Covering set).
Implementing a greedy algorithm to solve the NP-
Complete problem the covering set problem.